pyproject.toml
[tool.poetry]
name = "airtransportationdrf"
version = "0.1.0"
description = ""
authors = ["a1nsworth <xed255@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
black = "^23.12.0"
drf-yasg = "^1.21.7"
setuptools = "^69.0.2"
gunicorn = "^21.2.0"
environ = "^1.0"
dj-database-url = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
